REFACTOR: Decouple Start from Program
